+++ packages/gamut-styles/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Background';
+export * from './ColorMode';
+export * from './GamutProvider';
