+++ packages/gamut-styles/src/globals/index.tsx
@@ -1,2 +1,3 @@
 export * from './Typography';
 export * from './Reboot';
+export * from './Variables';
