+++ packages/gamut-styles/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useColorMode';
+export * from './useColorModes';
