+++ packages/gamut-styles/src/hooks/useColorMode.ts
@@ -1,0 +1,7 @@
+import { ColorModes } from '../types';
+import { useColorModes } from './useColorModes';
+
+export function useCurrentMode(mode?: ColorModes) {
+  const [activeMode] = useColorModes();
+  return mode ?? activeMode;
+}
