+++ packages/gamut-styles/src/index.ts
@@ -1,8 +1,6 @@
 import '@emotion/react';
 
-export * from './GamutProvider';
-export * from './ColorMode';
-export * from './Background';
+export * from './components';
 export * from './variance';
 export * from './cache';
 export * from './variables';
@@ -10,4 +8,5 @@ export * from './styles';
 export * from './globals';
 export * from './utilities';
 export * from './themes';
+export * from './types';
 export { coreTheme as theme } from './themes';
