+++ packages/gamut-styles/src/styles/modeColorProps.ts
@@ -1,0 +1,17 @@
+import { serializeTokens, ThemeProps } from '@codecademy/variance';
+import { mapValues } from 'lodash';
+
+import { ColorModes } from '../types';
+
+export const modeColorProps = ({
+  theme,
+  mode,
+}: ThemeProps<{ mode?: ColorModes }>) => {
+  if (!theme || !mode || mode === theme?.mode) return {};
+  const { colors } = theme;
+  return serializeTokens(
+    mapValues(theme?.modes[mode], (color) => colors[color]),
+    'color',
+    theme
+  ).variables;
+};
