+++ packages/gamut-styles/src/types.ts
@@ -1,0 +1,7 @@
+import { Theme } from '@emotion/react';
+
+export type Colors = keyof Theme['colors'];
+export type ColorModeConfig = Theme['modes'];
+export type ColorModes = keyof ColorModeConfig;
+export type ColorModeShape = ColorModeConfig[ColorModes];
+export type ColorAlias = keyof ColorModeShape;
